refactoring usecase layer
