+++ cmd/core/use-case/create-user/create-user.go
@@ -2,30 +2,22 @@ package createuserusecase
 
 import (
 	"go-api/cmd/core/entities/user"
+	portsservice "go-api/cmd/core/ports-service"
 	dto "go-api/cmd/interface/amqp/consumers/users/create/dto"
-
-	"github.com/google/uuid"
 )
 
+type createUserUseCase struct {
+	UserService portsservice.UserService
+}
+
+func New(us portsservice.UserService) CreateUserUseCase {
+	return &createUserUseCase{
+		UserService: us,
+	}
+}
+
 func (cuuc *createUserUseCase) CreateUser(dto dto.CreateDto) (*user.User, error) {
-
-	u, err := cuuc.RepositoryUser.GetByEmail(dto.Email)
-
-	if err != nil {
-		return nil, err
-	}
-
-	if u.ID != uuid.Nil {
-		return nil, user.ErrUserAlreadyExists
-	}
-
-	new_u, err := user.New(dto.Email, dto.Password, dto.Name)
-
-	if err != nil {
-		return nil, err
-	}
-
-	cuuc.RepositoryUser.Create(new_u)
+	new_u, err := cuuc.UserService.CreateUser(dto)
 
 	return new_u, err
 }
