+++ cmd/core/use-case/delete-user/delete-user.go
@@ -1,28 +1,21 @@
 package deleteuserusecase
 
 import (
-	entity "go-api/cmd/core/entities"
-	user "go-api/cmd/core/entities/user"
-	"log"
-
-	"github.com/google/uuid"
+	portsservice "go-api/cmd/core/ports-service"
 )
 
-func (duuc *deleteUserUseCase) DeleteUser(id string) error {
-	id_uuid := entity.ConvertId(id)
+type deleteUserUseCase struct {
+	UserService portsservice.UserService
+}
 
-	u, err := duuc.RepositoryUser.GetById(id_uuid)
-
-	if u.ID == uuid.Nil {
-		log.Default().Print("Not found user with id:" + id)
-		return user.ErrUserNotFound
+func New(us portsservice.UserService) DeleteUserUseCase {
+	return &deleteUserUseCase{
+		UserService: us,
 	}
+}
 
-	if err != nil {
-		return err
-	}
+func (duuc *deleteUserUseCase) DeleteUser(id string) (err error) {
+	duuc.UserService.DeleteUser(id)
 
-	duuc.RepositoryUser.DeleteById(id_uuid)
-
-	return nil
+	return
 }
