--- cmd/core/use-case/delete-user/main.go
@@ -1,15 +1,0 @@
-package deleteuserusecase
-
-import (
-	"go-api/cmd/core/ports"
-)
-
-type deleteUserUseCase struct {
-	RepositoryUser ports.UserRepository
-}
-
-func New(ur ports.UserRepository) DeleteUserUseCase {
-	return &deleteUserUseCase{
-		RepositoryUser: ur,
-	}
-}
