+++ cmd/core/use-case/list-user/list-users.go
@@ -1,40 +1,18 @@
 package listusersusecase
 
 import (
+	portsservice "go-api/cmd/core/ports-service"
 	response_jsonplaceholder "go-api/cmd/infra/integrations/http/jsonplaceholder/responses"
 	"log"
-	"reflect"
 )
 
-func (luuc *listUsersUseCase) ListUsers() {
-	ujs, err := luuc.UsersCache.Get("users")
+type listUsersUseCase struct {
+	UserService portsservice.UserService
+}
 
-	if err != nil {
-		log.Default().Print("###Error:Job failed, fail cache ###")
-		return
-	}
-
-	if reflect.DeepEqual(ujs, []response_jsonplaceholder.User{}) {
-		ujs, err := luuc.IntegrationJsonPlaceHolder.GetUsers()
-
-		if err != nil {
-			log.Default().Print("###Error:Job failed, fail integration ###")
-			return
-		}
-
-		if ujs == nil {
-			log.Fatalln("###Error: error for get user in cache and integration###")
-			return
-		}
-
-		luuc.UsersCache.Set("users", ujs, 100)
-		log.Default().Print("***Set users in cache***")
-
-		printUsers(ujs)
-	} else {
-		log.Default().Print("***Get users by cache***")
-
-		printUsers(ujs)
+func New(us portsservice.UserService) ListUsersUseCase {
+	return &listUsersUseCase{
+		UserService: us,
 	}
 }
 
@@ -42,4 +20,10 @@ func printUsers(ujs []response_jsonplaceholder.User) {
 	for _, u := range ujs {
 		log.Default().Print("-user:" + u.Username + "-email:" + u.Email + "-")
 	}
+}
+
+func (luuc *listUsersUseCase) ListUsers() {
+	ujs := luuc.UserService.ListUsers()
+
+	printUsers(ujs)
 }
