+++ cmd/infra/repositories/cache/users-jsonplaceholder/users-jsonplaceholder.interface.go
@@ -1,9 +1,9 @@
-package ports
+package usersjsonplaceholdercache
 
 import integration "go-api/cmd/infra/integrations/http/jsonplaceholder/responses"
 
 type (
-	UsersCache interface {
+	UsersJsonPlaceholderCache interface {
 		Set(key string, user []integration.User, time int)
 		Get(key string) ([]integration.User, error)
 	}
