+++ cmd/infra/repositories/cache/users-jsonplaceholder/users-jsonplaceholder_test.go
@@ -1,4 +1,4 @@
-package users_cache
+package usersjsonplaceholdercache
 
 import (
 	"encoding/json"
